+++ app/src/main/java/com/example/pdietetyczny/models/FoodItem.kt
@@ -3,9 +3,9 @@ package com.example.pdietetyczny.models
 data class FoodItem(
     val id: Int,
     val name: String,
-    val calories: Double,
-    val protein: Double,
-    val sugar: Double,
-    val fat: Double,
+    val calories: Float,
+    val protein: Float,
+    val sugar: Float,
+    val fat: Float,
 
 )
